feat(wip): add support for detection of `zstd` or `snappy` compression
Victoria Metrics support use of `zstd` compression, add support to
handle that distinction by using the NATS subject as a way to encode
this information. This seemed the simplist route to also still support
older versions versus without encapsulating the payload into a message
format.

None of this tested yet.